friend, like 오류수정

+++ src/main/java/org/example/newsfeedproject/like/service/CommentLikeService.java
@@ -5,8 +5,11 @@ import lombok.RequiredArgsConstructor;
 import org.example.newsfeedproject.comment.entity.Comment;
 import org.example.newsfeedproject.comment.repository.CommentRepository;
 import org.example.newsfeedproject.like.entity.CommentLike;
+import org.example.newsfeedproject.like.entity.FeedLike;
 import org.example.newsfeedproject.like.repository.CommentLikeRepository;
+import org.example.newsfeedproject.user.dto.SessionUserDto;
 import org.example.newsfeedproject.user.entity.User;
+import org.example.newsfeedproject.user.repository.UserRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -19,30 +22,43 @@ public class CommentLikeService {
 
     private final CommentLikeRepository commentLikeRepository;
     private final CommentRepository commentRepository;
+    private final UserRepository userRepository;
 
     public void commentLike(Long id, HttpServletRequest request) {
 
-        User user = (User) request.getSession().getAttribute("user");
+        SessionUserDto user = (SessionUserDto) request.getSession().getAttribute("user");
+        User me = userRepository.findByIdOrElseThrow(user.getId());
 
-        Comment comment = commentRepository.findByIdOrElseThrow(id);
+        Comment comment = commentRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 댓글을 찾을 수 없습니다."));
 
-        Optional<CommentLike> optionalCommentLike = commentLikeRepository.findByUserIdAndCommentId(user, comment);
+        if (me.getId().equals(comment.getUser().getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "본인의 댓글에 좋아요는 불가합니다.");
+        }
 
-        if (optionalCommentLike != null) {
+        CommentLike commentLike = commentLikeRepository.findByUserIdAndCommentId(me, comment);
+
+        if (commentLike != null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "중복 좋아요는 불가합니다.");
         }
 
-        commentLikeRepository.save(optionalCommentLike.get());
+        commentLike = new CommentLike(me, comment);
+
+        commentLikeRepository.save(commentLike);
 
     }
 
     public void deleteCommentLike(Long id, HttpServletRequest request) {
 
-        User user = (User) request.getSession().getAttribute("user");
+        SessionUserDto user = (SessionUserDto) request.getSession().getAttribute("user");
+        User me = userRepository.findByIdOrElseThrow(user.getId());
 
-        Comment comment = commentRepository.findByIdOrElseThrow(id);
+        Comment comment = commentRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 댓글을 찾을 수 없습니다."));
 
-        CommentLike commentLike = new CommentLike(user, comment);
+        CommentLike commentLike = commentLikeRepository.findByUserIdAndCommentId(me, comment);
+
+        if (commentLike == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "취소할 좋아요가 없습니다.");
+        }
 
         commentLikeRepository.delete(commentLike);
